# docs/swagger/openapi.yaml
...
openapi: 3.0.3
info:
  title: Foundations
  version: '1.0'
  description: The API for the Foundations project
servers:
  - url: http://localhost:8081
  - url: http://localhost:8080
paths:
...
  "/api/v1/transactions/{id}":
    get:
      security:
       [{
        bearerAuth: []
      }]
      tags:
      - Transaction
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Transaction"
            application/json:
              schema:
                "$ref": "#/components/schemas/Transaction"
            text/json:
              schema:
                "$ref": "#/components/schemas/Transaction"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    delete:
      security:
       [{
        bearerAuth: []
      }]
      tags:
      - Transaction
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Transaction"
            application/json:
              schema:
                "$ref": "#/components/schemas/Transaction"
            text/json:
              schema:
                "$ref": "#/components/schemas/Transaction"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
...
  "/api/v1/users/{id}":
    put:
      security:
       [{
        bearerAuth: []
      }]
      tags:
      - User
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserPost"
          text/json:
            schema:
              "$ref": "#/components/schemas/UserPost"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/UserPost"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/User"
            application/json:
              schema:
                "$ref": "#/components/schemas/User"
            text/json:
              schema:
                "$ref": "#/components/schemas/User"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '409':
          description: Conflict
    delete:
      security:
       [{
        bearerAuth: [],
      }]
      tags:
      - User
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/User"
            application/json:
              schema:
                "$ref": "#/components/schemas/User"
            text/json:
              schema:
                "$ref": "#/components/schemas/User"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    get:
      security:
       [{
        bearerAuth: []
      }]
      tags:
      - User
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/User"
            application/json:
              schema:
                "$ref": "#/components/schemas/User"
            text/json:
              schema:
                "$ref": "#/components/schemas/User"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
  "/api/v1/users/{id}/donate":
    post:
        security:
          [{
          bearerAuth: []
        }]
        tags:
        - User
        parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        requestBody:
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/UserDonate"
            text/json:
              schema:
                "$ref": "#/components/schemas/UserDonate"
            application/*+json:
              schema:
                "$ref": "#/components/schemas/UserDonate"
        responses:
          '200':
            description: Success
            content:
              text/plain:
                schema:
                  "$ref": "#/components/schemas/Transaction"
              application/json:
                schema:
                  "$ref": "#/components/schemas/Transaction"
              text/json:
                schema:
                  "$ref": "#/components/schemas/Transaction"
          '400':
            description: Bad Request
          '401':
            description: Unauthorized
          '404':
            description: Not Found
          '409':
            description: Conflict
...
components:
  schemas:
...
    User:
      type: object
      properties:
        Id:
          type: integer
          format: int32
        Login:
          type: string
        Password:
          type: string
        Balance:
          type: number
          format: double
        CharitySum:
          type: integer
          format: int32
        ClosedFingAmount:
          type: integer
          format: int32
      additionalProperties: false
    UserPost:
      type: object
      properties:
        Login:
          type: string
        Password:
          type: string
      additionalProperties: false
    UserDonate:
      type: object
      properties:
        entity-type:
          type: integer
          format: int32
        entity-id:
          type: integer
          format: int32
        sum_of_money:
          type: integer
          format: int32
        comm:
          type: string
      additionalProperties: false
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
          format: int32
          nullable: true
        fromId:
          type: integer
          format: int32
          nullable: true
        toId:
          type: integer
          format: int32
          nullable: true
        fromType:
          type: string
          nullable: true
        toType:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        sum:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
  securitySchemes:
      bearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT